Added a feature that allows user to save game.
This was done by adding a module to the program that provides general, basic serialization methods to the classes that include it.

Upon saving, the program serializes the the instance variables of the game object (the game data in this case), creates a save file, and stores the serialized string within the save file.

Upon execution of the program, if a save file is present, the user will be prompted if they want to load their previous game. The serialized string is loaded from the save file, deserialized, and a new game object is created with the save data, allowing the user to start where they left off.